+++ src-tauri/src/plugins/dmm/parsers/mod.rs
@@ -1,6 +1,6 @@
 //! DMM通販パーサーモジュール
 //!
-//! 各パーサーは dmm/ サブモジュールとして配置されています。
+//! 各パーサーは parsers/ サブモジュールとして配置されています。
 
 pub mod cancel;
 pub mod confirm;
