+++ src-tauri/src/plugins/hobbysearch/parsers/cancel.rs
