+++ src-tauri/src/plugins/hobbysearch/parsers/confirm.rs
