+++ src-tauri/src/plugins/hobbysearch/parsers/mod.rs
